Time: 4 ms (84.73%) | Memory: 10.9 MB (7.23%) - LeetSync

+++ 64-minimum-path-sum/minimum-path-sum.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int solve(int m, int n, vector<vector<int>> &grid, vector<vector<int>> &dp){
+    int solve(int m, int n, vector<vector<int>> &grid, vector<vector<long long>> &dp){
         if(m==0 && n==0){
             return grid[0][0];
         }
@@ -26,9 +26,9 @@ public:
         return dp[m][n] = min(up, left);
     }
     int minPathSum(vector<vector<int>>& grid) {
-        int m = grid.size();
-        int n = grid[0].size();
-        vector<vector<int>> dp(m, vector<int>(n, -1));
+        long long m = grid.size();
+        long long n = grid[0].size();
+        vector<vector<long long >> dp(m, vector<long long>(n, -1));
 
         return solve(m-1, n-1, grid, dp);
     }
